Fix handling of mixed text and CDATA nodes.
In XML elements can contain both text and CDATA nodes as children. The previous code would discard all but the last child node. hard-xml doesn't support accessing the different nodes directly so the most sensible thing to do is to merge the content together. This is inline with browser's `.textContent` attribute and the recommendation for encoding `]]>` in a CDATA sections as `<![CDATA[]]]]><![CDATA[>]]>` (aka closing the CDATA section and starting a new one mid-way through the forbidden sequence, so two sections containing `]]` and `>` respectively.).

https://www.w3.org/TR/REC-xml/#d0e2433